delete DashboardMetadataDAO interface & rename DashboardMetadataDAOImpl
to DashboardMetadataDAO
